+++ spec/javascripts/govuk_publishing_components/analytics-ga4/pii-remover.spec.js
@@ -14,21 +14,23 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
 
   describe('by default', function () {
     it('strips email addresses, but not postcodes and dates from strings', function () {
-      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date')
-      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a 2019-01-21 date')
+      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date, and this is another 1 January 1990 date')
+      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a 2019-01-21 date, and this is another 1 January 1990 date')
     })
 
     it('strips email addresses but not dates and postcodes from objects', function () {
       var obj = {
         email: 'this is an@email.com address',
         postcode: 'this is a sw1a 1aa postcode',
-        date: 'this is a 2019-01-21 date'
+        date: 'this is a 2019-01-21 date',
+        date2: 'this is another 1 January 1990 date'
       }
 
       var strippedObj = {
         email: 'this is [email] address',
         postcode: 'this is a sw1a 1aa postcode',
-        date: 'this is a 2019-01-21 date'
+        date: 'this is a 2019-01-21 date',
+        date2: 'this is another 1 January 1990 date'
       }
 
       obj = pii.stripPII(obj)
@@ -39,13 +41,15 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
       var arr = [
         'this is an@email.com address',
         'this is a sw1a 1aa postcode',
-        'this is a 2019-01-21 date'
+        'this is a 2019-01-21 date',
+        'this is another 1 January 1990 date'
       ]
 
       var strippedArr = [
         'this is [email] address',
         'this is a sw1a 1aa postcode',
-        'this is a 2019-01-21 date'
+        'this is a 2019-01-21 date',
+        'this is another 1 January 1990 date'
       ]
 
       arr = pii.stripPII(arr)
@@ -74,8 +78,8 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
     })
 
     it('strips email addresses, postcodes and dates from strings', function () {
-      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date')
-      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a [date] date')
+      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date, and this is another 1 January 1990 date')
+      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a [date] date, and this is another [date] date')
     })
 
     it('strips all PII from objects', function () {
@@ -83,6 +87,7 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
         email: 'this is an@email.com address',
         postcode: 'this is a sw1a 1aa postcode',
         date: 'this is a 2019-01-21 date',
+        date2: 'this is another 1 January 1990 date',
         uuid: 'd6c2de5d-ef90-45d1-82d4-5f2438369eea'
       }
 
@@ -90,6 +95,7 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
         email: 'this is [email] address',
         postcode: 'this is a [postcode] postcode',
         date: 'this is a [date] date',
+        date2: 'this is another [date] date',
         uuid: 'd6c2de5d-ef90-45d1-82d4-5f2438369eea'
       }
 
@@ -101,13 +107,15 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
       var arr = [
         'this is an@email.com address',
         'this is a sw1a 1aa postcode',
-        'this is a 2019-01-21 date'
+        'this is a 2019-01-21 date',
+        'and this is another 1 January 1990 date'
       ]
 
       var strippedArr = [
         'this is [email] address',
         'this is a [postcode] postcode',
-        'this is a [date] date'
+        'this is a [date] date',
+        'and this is another [date] date'
       ]
 
       arr = pii.stripPII(arr)
@@ -123,8 +131,8 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
 
     it('observes the meta tag and strips out postcodes', function () {
       expect(pii.stripPostcodePII).toEqual(true)
-      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a long GL194LYX postcode, this is a 2019-01-21 date, this is a d6c2de5d-ef90-45d1-82d4-5f2438369eea content ID, this is a p800refund')
-      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a long [postcode] postcode, this is a 2019-01-21 date, this is a d6c2de5d-ef90-45d1-82d4-5f2438369eea content ID, this is a p800refund')
+      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a long GL194LYX postcode, this is a 2019-01-21 date, this is another 1 January 1990 date, this is a d6c2de5d-ef90-45d1-82d4-5f2438369eea content ID, this is a p800refund')
+      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a long [postcode] postcode, this is a 2019-01-21 date, this is another 1 January 1990 date, this is a d6c2de5d-ef90-45d1-82d4-5f2438369eea content ID, this is a p800refund')
     })
 
     it('doesn\'t strip out UUIDs in URLs', function () {
@@ -172,8 +180,8 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
 
     it('observes the meta tag and strips out dates', function () {
       expect(pii.stripDatePII).toEqual(true)
-      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date')
-      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a [date] date')
+      var string = pii.stripPII('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date, and this is another 1 January 1990 date')
+      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a [date] date, and this is another [date] date')
     })
   })
 
@@ -185,23 +193,23 @@ describe('GOVUK.analyticsGa4.PIIRemover', function () {
     })
 
     it('observes the override and strips out emails', function () {
-      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date', false, false)
-      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a 2019-01-21 date')
+      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date and this is another 1 January 1990 date', false, false)
+      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a 2019-01-21 date and this is another 1 January 1990 date')
     })
 
     it('observes the override and strips out emails and dates', function () {
-      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date', true, false)
-      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a [date] date')
+      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date and this is another 1 January 1990 date', true, false)
+      expect(string).toEqual('this is [email] address, this is a sw1a 1aa postcode, this is a [date] date and this is another [date] date')
     })
 
     it('observes the override and strips out emails and postcodes', function () {
-      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date', false, true)
-      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a 2019-01-21 date')
+      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date and this is another 1 January 1990 date', false, true)
+      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a 2019-01-21 date and this is another 1 January 1990 date')
     })
 
     it('observes the override and strips out emails, postcodes and dates', function () {
-      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date', true, true)
-      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a [date] date')
+      var string = pii.stripPIIWithOverride('this is an@email.com address, this is a sw1a 1aa postcode, this is a 2019-01-21 date and this is another 1 January 1990 date', true, true)
+      expect(string).toEqual('this is [email] address, this is a [postcode] postcode, this is a [date] date and this is another [date] date')
     })
   })