+++ spec/components/tabs_spec.rb
@@ -24,10 +24,38 @@ describe "Tabs", type: :view do
         },
       ],
     )
-    assert_select ".govuk-tabs"
+    assert_select ".govuk-tabs[data-module='govuk-tabs']"
     assert_select ".govuk-tabs__tab", 2
     assert_select ".govuk-tabs__panel", 2
     assert_select "#tab1"
     assert_select "#tab1"
+  end
+
+  it "renders with GA4 tracking" do
+    render_component(
+      ga4_tracking: true,
+      tabs: [
+        {
+          id: "tab1",
+          label: "First section",
+          content: "<p>Fusce at dictum tellus, ac accumsan est. Nulla vitae turpis in nulla gravida tincidunt.</p>",
+        },
+        {
+          id: "tab2",
+          label: "Second section",
+          content: "<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>",
+        },
+        {
+          id: "tab3",
+          label: "Third section",
+          content: "<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>",
+        },
+      ],
+    )
+
+    assert_select ".govuk-tabs[data-module='govuk-tabs ga4-event-tracker']"
+    assert_select ".govuk-tabs__tab[data-ga4-event='{\"event_name\":\"select_content\",\"type\":\"tabs\",\"text\":\"First section\",\"index\":{\"index_section\":1,\"index_section_count\":3}}']"
+    assert_select ".govuk-tabs__tab[data-ga4-event='{\"event_name\":\"select_content\",\"type\":\"tabs\",\"text\":\"Second section\",\"index\":{\"index_section\":2,\"index_section_count\":3}}']"
+    assert_select ".govuk-tabs__tab[data-ga4-event='{\"event_name\":\"select_content\",\"type\":\"tabs\",\"text\":\"Third section\",\"index\":{\"index_section\":3,\"index_section_count\":3}}']"
   end
 end